server/pyproject.toml
[tool.poetry]
name = "twitter-django"
version = "0.1.0"
description = ""
authors = ["Kevin Cunningham <kevin@kevincunningham.co.uk>"]
readme = "README.md"
packages = [{include = "twitter_django"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.5"
django-rest-framework = "^0.1.0"
django-cors = "^0.1"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.5"
pyjwt = "^2.6.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
